# src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sanktionssök API",
    "description": "REST API för att söka mot konsoliderade sanktionslistor (EU, FN, US OFAC) samt PEP-listor.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Lokal utvecklingsserver"
    }
  ],
  "paths": {
...
    "/api/overrides/{id}": {
      "put": {
        "summary": "Skapa eller ersätta en override för en sanktionspost",
        "description": "Lägger en fältvis rättelse ovanpå en importerad post (issue #35). Den lagrade posten muteras aldrig — tar man bort overriden kommer de senast importerade värdena tillbaka exakt. Ersätter (upsert), mergar inte fältvis med en tidigare override. Kräver inloggad session; overriddenBy sätts alltid från den inloggade användaren, aldrig från request-body.",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID för sanktionsposten som ska overridas (t.ex. EU-1234)",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fields", "reason"],
                "properties": {
                  "fields": {
                    "type": "object",
                    "description": "Icke-tomt objekt med fält att rätta. id/source/type/createdAt/searchNames/status kan inte overridas.",
                    "example": { "sanctionReason": "Korrigerad orsakstext" }
                  },
                  "reason": {
                    "type": "string",
                    "example": "Källans transliterering var felaktig"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Override sparad",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Override" }
              }
            }
          },
          "400": {
            "description": "Saknat/tomt \"fields\" eller \"reason\", eller försök att overrida ett skyddat fält"
          },
          "401": {
            "description": "Ingen giltig session (ej inloggad)"
          },
          "404": {
            "description": "Ingen sanktionspost med detta ID hittades"
          }
        }
      },
      "delete": {
        "summary": "Ta bort en override",
        "description": "Återställer exakt de senast importerade värdena för posten — den lagrade posten muterades aldrig av overriden.",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Override borttagen (idempotent — 200 även om ingen fanns)",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } }
              }
            }
          },
          "401": {
            "description": "Ingen giltig session (ej inloggad)"
          }
        }
      }
    },
...
    "/api/admin/custom-records": {
      "post": {
        "summary": "Skapa en intern watchlist-post (source: CUSTOM)",
        "description": "Skapar en manuellt tillagd sanktionspost (t.ex. en lokal PEP eller intern bevakningspost) som inte täcks av de officiella källorna (issue #172). Existens-kontrollen sker atomärt via Firestores DocumentReference.create() — två samtidiga skapanden med samma id kan aldrig båda lyckas. Kräver en inloggad admin-session.",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id", "type", "primaryName"],
                "properties": {
                  "id": { "type": "string", "example": "CUSTOM-1" },
                  "type": { "type": "string", "enum": ["individual", "entity", "vessel"] },
                  "primaryName": { "type": "string", "example": "Jane Doe" },
                  "aliases": { "type": "array", "items": { "type": "string" } },
                  "datesOfBirth": { "type": "array", "items": { "type": "string" } },
                  "placesOfBirth": { "type": "array", "items": { "type": "string" } },
                  "citizenships": { "type": "array", "items": { "type": "string" } },
                  "passports": { "type": "array", "items": { "type": "string" } },
                  "sanctionReason": { "type": "string" },
                  "legalBasis": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Posten skapad",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SanctionRecord" }
              }
            }
          },
          "400": {
            "description": "Saknat/ogiltigt fält, t.ex. tomt primaryName eller ett id med otillåtna tecken"
          },
          "401": {
            "description": "Ingen giltig session (ej inloggad)"
          },
          "403": {
            "description": "Inloggad men inte admin"
          },
          "409": {
            "description": "En post med detta id finns redan"
          }
        }
      }
    },
    "/api/admin/custom-records/{id}": {
      "put": {
        "summary": "Uppdatera en intern watchlist-post",
        "description": "Patchar fält på en befintlig CUSTOM-post. Vägrar röra en post med annan source (använd overrides-vägen för sådana). id/source/createdAt/searchNames pinnas alltid om efter patchen — searchNames genereras alltid på nytt från den sammanslagna postens eget primaryName/aliases, aldrig från ett värde i request body.",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Valfri delmängd av samma fält som POST accepterar",
                "example": { "sanctionReason": "Uppdaterad intern notering" }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Posten uppdaterad",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SanctionRecord" }
              }
            }
          },
          "400": {
            "description": "Ogiltigt id, eller posten är inte en CUSTOM-post"
          },
          "401": {
            "description": "Ingen giltig session (ej inloggad)"
          },
          "403": {
            "description": "Inloggad men inte admin"
          },
          "404": {
            "description": "Ingen CUSTOM-post med detta id hittades"
          }
        }
      },
      "delete": {
        "summary": "Ta bort en intern watchlist-post",
        "description": "Kräver explicit confirm: true i request body — annars 400. Vägrar radera en post med annan source än CUSTOM.",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["confirm"],
                "properties": { "confirm": { "type": "boolean", "example": true } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Posten borttagen",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } }
              }
            }
          },
          "400": {
            "description": "confirm: true saknas, eller posten är inte en CUSTOM-post"
          },
          "401": {
            "description": "Ingen giltig session (ej inloggad)"
          },
          "403": {
            "description": "Inloggad men inte admin"
          },
          "404": {
            "description": "Ingen CUSTOM-post med detta id hittades"
          }
        }
      },
      "get": {
        "summary": "Hämta en intern watchlist-post",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Posten",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SanctionRecord" }
              }
            }
          },
          "401": {
            "description": "Ingen giltig session (ej inloggad)"
          },
          "403": {
            "description": "Inloggad men inte admin"
          },
          "404": {
            "description": "Ingen post med detta id hittades"
          }
        }
      }
    },
    "/api/decisions": {
      "post": {
        "summary": "Registrera eller uppdatera en screening-adjudikation",
        "description": "Skapar eller skriver över (upsert) den aktuella false-positive/true-positive-bedömningen för ett entity+subject-par. `decidedBy` sätts alltid från den inloggade sessionen, aldrig från request body. Att adjudikera på nytt (t.ex. en annan analytiker som är oense) skriver bara över tidigare bedömning för samma par — det är normalfallet, inte ett fel (issue #22).",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["entityId", "subjectId", "verdict"],
                "properties": {
                  "entityId": { "type": "string", "example": "EU-1234" },
                  "subjectId": { "type": "string", "example": "customer-acme" },
                  "verdict": { "type": "string", "enum": ["false_positive", "true_positive"] },
                  "notes": { "type": "string", "example": "Confirmed match on passport number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bedömning sparad",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Decision" }
              }
            }
          },
          "400": {
            "description": "Ogiltig entityId/subjectId/verdict, eller decidedBy saknas"
          },
          "401": {
            "description": "Ingen giltig session (ej inloggad)"
          }
        }
      }
    },
    "/api/decisions/{entityId}": {
      "get": {
        "summary": "Lista alla bedömningar för en entity",
        "description": "Returnerar varje sparad bedömning för entityn, över alla subjects.",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista över bedömningar",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Decision" }
                }
              }
            }
          },
          "401": {
            "description": "Ingen giltig session (ej inloggad)"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__session",
        "description": "Session-cookie satt av POST /api/auth/verify-otp efter lyckad e-post-OTP-inloggning. Krävs av alla /api/*-rutter förutom auth-rutterna själva och /api/admin/tokens/* (endast admin-session, se requireAdmin)."
      },
      "apiTokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "sanc_<random>",
        "description": "API-token skapad via POST /api/admin/tokens, skickad som 'Authorization: Bearer sanc_...'. Stödjer granulära scopes (sanctions:read, custom:read/write, overrides:read/write, decisions:read/write, imports:read/write, system:read) samt bakåtkompatibla 'read'/'write'-paket (issue #36, #221)."
      }
    },
    "schemas": {
...
      "Decision": {
        "type": "object",
        "properties": {
          "entityId": { "type": "string", "example": "EU-1234" },
          "subjectId": { "type": "string", "example": "customer-acme" },
          "verdict": { "type": "string", "enum": ["false_positive", "true_positive"] },
          "decidedBy": { "type": "string", "example": "analyst@example.com" },
          "decidedAt": { "type": "string", "format": "date-time" },
          "notes": { "type": "string", "nullable": true }
        }
      },
...
      "SanctionRecord": {
        "type": "object",
        "required": ["id", "source", "type", "names", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "example": "EU-1234"
          },
          "source": {
            "type": "string",
            "enum": ["EU", "UN", "US", "UK", "PEP", "CUSTOM"],
            "example": "EU"
          },
          "type": {
            "type": "string",
            "enum": ["individual", "entity", "vessel", "aircraft"],
            "example": "individual"
          },
          "names": {
            "type": "array",
            "description": "Every name/alias for this entity, primary name first (issue #46 — the sole source of truth; no separate flat primaryName/aliases fields).",
            "items": {
              "type": "object",
              "required": ["wholeName", "strong"],
              "properties": {
                "wholeName": { "type": "string", "example": "Vladimir Vladimirovich Putin" },
                "firstName": { "type": "string" },
                "middleName": { "type": "string" },
                "lastName": { "type": "string" },
                "strong": { "type": "boolean", "description": "Authoritative name vs. a weaker variant, per the source." },
                "language": { "type": "string" },
                "title": { "type": "string" },
                "function": { "type": "string" }
              }
            },
            "example": [
              { "wholeName": "Vladimir Vladimirovich Putin", "strong": true },
              { "wholeName": "Valodia Putin", "strong": false }
            ]
          },
          "birthDates": {
            "type": "array",
            "description": "Structured, precision-aware birth dates (exact date, year-only, year range, or circa/approximate).",
            "items": {
              "type": "object",
              "properties": {
                "raw": { "type": "string", "example": "1952-10-07" },
                "year": { "type": "integer" },
                "month": { "type": "integer" },
                "day": { "type": "integer" },
                "yearRangeFrom": { "type": "integer" },
                "yearRangeTo": { "type": "integer" },
                "circa": { "type": "boolean" },
                "place": { "type": "string" },
                "city": { "type": "string" },
                "countryIso2": { "type": "string" }
              }
            }
          },
          "placesOfBirth": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Leningrad, USSR"]
          },
          "citizenships": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Russia"]
          },
          "identifications": {
            "type": "array",
            "description": "Passports, national IDs, and similar documents — structured, including the source's own reliability flags (CLAUDE.md §6: never present a known-false/expired/revoked document as good data).",
            "items": {
              "type": "object",
              "required": ["number"],
              "properties": {
                "number": { "type": "string", "example": "123456" },
                "typeCode": { "type": "string" },
                "typeDescription": { "type": "string", "example": "Diplomatic Passport" },
                "countryIso2": { "type": "string" },
                "issuedBy": { "type": "string" },
                "knownFalse": { "type": "boolean" },
                "knownExpired": { "type": "boolean" },
                "reportedLost": { "type": "boolean" },
                "revokedByIssuer": { "type": "boolean" },
                "diplomatic": { "type": "boolean" }
              }
            }
          },
          "addresses": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "street": { "type": "string" },
                "city": { "type": "string" },
                "country": { "type": "string" },
                "fullAddress": { "type": "string" }
              }
            }
          },
          "sanctionReason": {
            "type": "string",
            "example": "Involvement in destabilizing Ukraine"
          },
          "legalBasis": {
            "type": "string",
            "example": "Regulation (EU) No 269/2014"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": ["active", "delisted"],
            "description": "Soft-delete status. A missing value should be treated as 'active'.",
            "example": "active"
          },
          "listedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record first appeared in an import."
          },
          "delistedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record most recently went missing from an import."
          },
          "overriddenFields": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Field names currently overridden by a local correction (issue #35). Not part of the persisted document — computed on every read by merging overrides/{id} onto the stored record, so the caller can tell official source data apart from a local correction. Empty when no override exists for this record.",
            "example": ["sanctionReason"]
          }
        }
      },
      "Override": {
        "type": "object",
        "required": ["entityId", "fields", "overriddenBy", "overriddenAt", "reason"],
        "properties": {
          "entityId": { "type": "string", "example": "EU-1234" },
          "fields": {
            "type": "object",
            "description": "Sparse set of fields to override. Immutable identity/bookkeeping fields (id, source, type, createdAt, searchNames, status) are rejected — status cannot be overridden so a correction can never resurrect a delisted record.",
            "additionalProperties": true
          },
          "overriddenBy": {
            "type": "string",
            "description": "Always the authenticated caller's email — never taken from the request body.",
            "example": "analyst@example.com"
          },
          "overriddenAt": { "type": "string", "format": "date-time" },
          "reason": { "type": "string", "example": "Source transliteration was inaccurate" }
        }
      },
...
    }
  }
}